Handle lines with a property name and no value.
This commit fixes a defect in the handling of lines with only a property name, and maybe a separator, but no value.  In those cases the property is considered set to an empty string.  (Prior to this change, such cases would set the property name to the contents of the next non-empty line, even if it was a comment.)